Cant delete book when book is loaned

+++ web/src/main/java/DeleteServlet.java
@@ -1,5 +1,8 @@
+import dto.BookDto;
 import repositories.BooksRepository;
 import repositories.IBooksRepository;
+import services.BookService;
+import services.IBookService;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -11,16 +14,25 @@ import java.io.IOException;
 @WebServlet(urlPatterns = "/DeleteServlet")
 public class DeleteServlet extends HttpServlet {
     private IBooksRepository repository = new BooksRepository();
+    private final IBookService service;
+
+    public DeleteServlet() {
+        this.service = new BookService();
+    }
+
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 
     }
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 
-            String bookId = request.getParameter("bookId");
-            bookId = bookId.trim();
+        String bookId = request.getParameter("bookId");
+        bookId = bookId.trim();
+        BookDto bookDto = service.findBook(Integer.valueOf(bookId));
+        if (bookDto.isBorrowed()) {
             repository.delete(Integer.valueOf(bookId));
-            response.sendRedirect("/HomeServlet");
+        }
+        response.sendRedirect("/HomeServlet");
 
     }
 }
